+++ supportedBackends/jes/src/main/scala/cromwell/backend/impl/jes/JesBackendSingletonActor.scala
@@ -1,12 +1,12 @@
 package cromwell.backend.impl.jes
 
 import akka.actor.{Actor, ActorLogging, Props}
-import cromwell.backend.impl.jes.statuspolling.{JesApiQueryManager}
+import cromwell.backend.impl.jes.statuspolling.JesApiQueryManager
 import cromwell.backend.impl.jes.statuspolling.JesApiQueryManager.DoPoll
 
-class JesBackendSingletonActor extends Actor with ActorLogging {
+final case class JesBackendSingletonActor(qps: Int) extends Actor with ActorLogging {
 
-  val pollingActor = context.actorOf(JesApiQueryManager.props)
+  val pollingActor = context.actorOf(JesApiQueryManager.props(qps))
 
   override def receive = {
     case poll: DoPoll =>
@@ -16,5 +16,5 @@ class JesBackendSingletonActor extends Actor with ActorLogging {
 }
 
 object JesBackendSingletonActor {
-  def props(): Props = Props(new JesBackendSingletonActor())
+  def props(qps: Int): Props = Props(JesBackendSingletonActor(qps))
 }